refactor: track upstream module changes (#835)
